+++ hiqlite/src/start.rs
@@ -105,6 +105,7 @@ where
         #[cfg(any(feature = "backup", feature = "dashboard"))]
         tx_client_stream: tx_client_stream.clone(),
         health_check_delay_secs: node_config.health_check_delay_secs,
+        learner_only: node_config.learner_only,
         #[cfg(feature = "s3")]
         s3_config: node_config.s3_config.clone(),
     });